In computer class, the first assignment was to write a program to print the first 100 Fibonacci numbers. Instead, I wrote a program that would steal passwords of students. My teacher gave me an A+

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -9,19 +9,19 @@
 
 int main(void)
 {
-	long long int i, a, b;
+	unsigned long int i, a, b;
 
 	a = 1;
 	b = 2;
 
-	printf("%lld, %lld, ", a, b);
+	printf("%ld, %ld, ", a, b);
 
 	for (i = 2; i < 98; i++)
 	{
 		b = a + b;
 		a = b - a;
 
-		printf("%lld", b);
+		printf("%ld", b);
 
 		if (i < 97)
 			printf(", ");
